refactor(plugins): separate consent mutation from inventory
