+++ src/plugins/capability-consent.ts
@@ -1,10 +1,8 @@
-import { createHash } from "node:crypto";
 import fs from "node:fs";
 import path from "node:path";
 import { redactSensitiveUrlLikeString } from "@openclaw/net-policy/redact-sensitive-url";
 import { PLUGIN_DECLARED_SURFACE_GROUPS } from "../../packages/gateway-protocol/src/schema/plugin-declared-surface-groups.js";
 import type {
-  PluginInspectSource,
   PluginInstallTrust,
   PluginsInspectResult,
 } from "../../packages/gateway-protocol/src/schema/plugins.js";
@@ -24,7 +22,10 @@ import { isPathInside } from "../infra/path-guards.js";
 import { resolveUserPath } from "../utils.js";
 import {
   buildPluginCapabilitySummary,
+  computeDeclaredSurfaceHash,
   mergePluginDeclaredSurfaces,
+  resolveAcceptedSurfaceCurrent,
+  resolvePluginInstallRecordIntegrity,
   resolvePluginPackageDeclaredSurface,
 } from "./capability-summary.js";
 import { resolvePluginControlPlaneWorkspace } from "./control-plane-workspace.js";
@@ -174,13 +175,6 @@ export function resolvePluginArtifactDeclaredSurface(
   );
 }
 
-export function computeDeclaredSurfaceHash(declared: PluginAcceptedDeclaredSurface): string {
-  const canonical = Object.fromEntries(
-    PLUGIN_DECLARED_SURFACE_GROUPS.map((group) => [group, declared[group].toSorted()]),
-  );
-  return createHash("sha256").update(JSON.stringify(canonical)).digest("hex");
-}
-
 export function diffDeclaredSurfaceWidening(
   previous: PluginAcceptedDeclaredSurface,
   next: PluginAcceptedDeclaredSurface,
@@ -194,36 +188,6 @@ export function diffDeclaredSurfaceWidening(
     }
   }
   return { widened, hasWidening: Object.keys(widened).length > 0 };
-}
-
-export function resolvePluginInstallRecordIntegrity(
-  record: Pick<PluginInstallRecord, "integrity" | "npmIntegrity" | "clawpackSha256" | "gitCommit">,
-):
-  | { integrity: string; integrityKind: NonNullable<PluginInspectSource["integrityKind"]> }
-  | undefined {
-  const npmIntegrity = record.integrity ?? record.npmIntegrity;
-  if (npmIntegrity) {
-    return { integrity: npmIntegrity, integrityKind: "ssri" };
-  }
-  if (record.clawpackSha256) {
-    return { integrity: record.clawpackSha256, integrityKind: "sha256" };
-  }
-  return record.gitCommit
-    ? { integrity: record.gitCommit, integrityKind: "git-commit" }
-    : undefined;
-}
-
-export function resolveAcceptedSurfaceCurrent(
-  record: PluginInstallRecord,
-  declared: PluginAcceptedDeclaredSurface,
-): boolean {
-  return (
-    record.acceptedSurface !== undefined &&
-    record.acceptedSurfaceHash !== undefined &&
-    record.acceptedSurfaceHash === computeDeclaredSurfaceHash(record.acceptedSurface) &&
-    record.acceptedSurfaceHash === computeDeclaredSurfaceHash(declared) &&
-    record.acceptedSurfaceIntegrity === resolvePluginInstallRecordIntegrity(record)?.integrity
-  );
 }
 
 export type PluginCapabilityConsentAcknowledgment = { reviewToken: string };
@@ -675,8 +639,4 @@ export async function prepareManagedPluginArtifactConsentHandler(
     previousRecords,
     previousPluginOwners,
   });
-}
-
-export function formatPluginCapabilityConsentRequired(pluginId: string): string {
-  return `Plugin "${pluginId}" requires capability consent; disable and re-enable it or run \`openclaw plugins enable ${pluginId} --accept-capabilities\`.`;
 }
